backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pathclaw"
version = "0.1.0"
description = "Computational pathology platform with MIL/MAMMOTH training backend"
readme = {file = "../README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Devansh Lalwani", email = "devansh@turocrates.ai" },
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "mammoth-moe>=0.1.0",
    "openslide-python>=1.3.0",
    "Pillow>=10.0",
    "numpy>=1.24",
    "scikit-image>=0.21",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "einops>=0.7",
    "huggingface-hub>=0.20",
    "timm>=0.9",
    "websockets>=12.0",
    "aiofiles>=23.0",
    "python-multipart>=0.0.6",
    "httpx>=0.24",
    "peft>=0.7.0",
    "cellpose>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
pathclaw = "pathclaw.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"
